requirements.txt
numpy==1.20.1
prometheus_client==0.9.0
streamlit==0.76.0
pandas==1.2.2
Pillow==8.1.0
